control-panel: restore persisted repeat config
To let user-changed "enabled" settings for control-panel repeat commands
take precedence over those command's autostart settings and
default-enabled definitions, always convert the enabled_repeats value to
a boolean and pass it as the "enabled" argument of apply_command when
restoring repeat commands in apply_fort_loaded_config.

Because of a limitation in the persistence data (only the enabled repeat
commands are recorded), we can not readily tell the difference between a
"new" default-enabled repeat command (e.g. introduced in a newly updated
DFHack installation) and a previously existing repeat command that the
user intentionally disabled. This means that we can not automatically
enable such a "new" default-enabled repeat command when loading a DF
save under an updated DFHack installation. The user can still manually
enable such a "new" repeat command, and that setting will be persisted
and restored properly (new embarks are not affected since they are
handled in apply_autostart_config where the "enabled" argument is left
off to get the defaulting behavior).